--- src/main/java/com/intouch/aligooligo/User/Controller/Dto/UserLoginRequestDto.java
@@ -1,11 +1,0 @@
-package com.intouch.aligooligo.User.Controller.Dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserLoginRequestDto {
-    private String email;
-    private String password;
-}
