--- src/main/java/com/intouch/aligooligo/auth/dto/TokenResponse.java
@@ -1,11 +1,0 @@
-package com.intouch.aligooligo.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class TokenResponse {
-    private String accessToken;
-    private String expiredTime;
-}
